requirements.txt
aiohttp==3.7.4.post0
asyncpg
colorama==0.4.4
git+https://github.com/Rapptz/discord.py
humanfriendly==10.0
python-dotenv==0.20.0
nudepy==0.5.1
akinator==1.0.1
Wikipedia-API==0.5.4
